Added the new greedy strategy method with initial load, tests are pending.
