+++ src/main.cpp
@@ -1,5 +1,5 @@
 #include <lb/input/naiveBasicInputInitialLoad.h>
-#include <lb/impl/greedyLB.h>
+#include <lb/impl/greedyStrategy.h>
 
 /**
  * This main is temporary and only usable to probe for compilation errors as the third-party makefile only compiles the files which are dependencies of the main file.
